cleanup: optimize common part-2
