+++ fss/tests/system/v1/role_test.py
@@ -2,7 +2,7 @@ import pytest
 from fastapi.testclient import TestClient
 
 from fss.common.config import configs
-from fss.common.util.security import get_user_id
+from fss.common.security.security import get_user_id
 from fss.starter.server import app
 
 client = TestClient(app)
